+++ src/main/java/org/edu/ntnu/idatt2003/group49/millions/model/Player.java
@@ -1,4 +1,7 @@
-package org.edu.ntnu.idatt2003.group49.millions;
+package org.edu.ntnu.idatt2003.group49.millions.model;
+
+import org.edu.ntnu.idatt2003.group49.millions.Portfolio;
+import org.edu.ntnu.idatt2003.group49.millions.TransactionArchive;
 
 import java.math.BigDecimal;
 import java.util.Objects;
