+++ src/main/java/org/edu/ntnu/idatt2003/group49/millions/model/Stock.java
@@ -1,4 +1,4 @@
-package org.edu.ntnu.idatt2003.group49.millions;
+package org.edu.ntnu.idatt2003.group49.millions.model;
 
 import java.math.BigDecimal;
 import java.util.ArrayList;
