+++ src/test/java/org/edu/ntnu/idatt2003/group49/millions/calculator/SaleCalculatorTest.java
@@ -1,7 +1,7 @@
 package org.edu.ntnu.idatt2003.group49.millions.calculator;
 
-import org.edu.ntnu.idatt2003.group49.millions.Share;
-import org.edu.ntnu.idatt2003.group49.millions.Stock;
+import org.edu.ntnu.idatt2003.group49.millions.model.Share;
+import org.edu.ntnu.idatt2003.group49.millions.model.Stock;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
